+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## [v1.18.2](https://github.com/vazco/uniforms/tree/v1.18.1) (2017-06-09)
 - **Fixed:** `SubmitField` respects `disabled` prop.
 
 ## [v1.18.1](https://github.com/vazco/uniforms/tree/v1.18.1) (2017-05-28)